Move functions into folders flux and sky_spectrum, add import to __init__.py
